Extract customer select into Backbone View
This encapsulates the details of using select2
